+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/mangaraw/MangaRawGenerator.kt
@@ -8,11 +8,12 @@ class MangaRawGenerator : ThemeSourceGenerator {
 
     override val themePkg = "mangaraw"
 
-    override val baseVersionCode = 2
+    override val baseVersionCode = 4
 
     override val sources = listOf(
-        SingleLang("SyoSetu", "https://syosetu.top", "ja", overrideVersionCode = 1),
-        SingleLang("MangaRaw", "https://manga9.co", "ja", pkgName = "manga9co", overrideVersionCode = 1),
+        SingleLang("SyoSetu", "https://syosetu.top", "ja"),
+        SingleLang("MangaRaw", "https://manga1001.in", "ja", pkgName = "manga9co", overrideVersionCode = 2),
+        SingleLang("MangaRawRU", "https://mangaraw.ru", "ja", overrideVersionCode = 1),
     )
 
     companion object {