+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/mangathemesia/MangaThemesia.kt
@@ -30,12 +30,13 @@ import java.text.SimpleDateFormat
 import java.util.Locale
 import java.util.concurrent.TimeUnit
 
+// Formerly WPMangaStream & WPMangaReader -> MangaThemesia
 abstract class MangaThemesia(
     override val name: String,
     override val baseUrl: String,
     override val lang: String,
     val mangaUrlDirectory: String = "/manga",
-    private val dateFormat: SimpleDateFormat = SimpleDateFormat("MMMM dd, yyyy", Locale.US)
+    val dateFormat: SimpleDateFormat = SimpleDateFormat("MMMM dd, yyyy", Locale.US)
 ) : ParsedHttpSource() {
 
     protected open val json: Json by injectLazy()
@@ -188,6 +189,7 @@ abstract class MangaThemesia(
     open fun String?.parseStatus(): Int = when {
         this == null -> SManga.UNKNOWN
         listOf("ongoing", "publishing").any { this.contains(it, ignoreCase = true) } -> SManga.ONGOING
+        this.contains("hiatus", ignoreCase = true) -> SManga.ON_HIATUS
         this.contains("completed", ignoreCase = true) -> SManga.COMPLETED
         else -> SManga.UNKNOWN
     }
@@ -303,9 +305,9 @@ abstract class MangaThemesia(
     }
 
     // Filters
-    private class AuthorFilter : Filter.Text("Author")
+    protected class AuthorFilter : Filter.Text("Author")
 
-    private class YearFilter : Filter.Text("Year")
+    protected class YearFilter : Filter.Text("Year")
 
     open class SelectFilter(
         displayName: String,
@@ -362,7 +364,12 @@ abstract class MangaThemesia(
         )
     )
 
-    protected class Genre(name: String, val value: String) : Filter.TriState(name)
+    protected class Genre(
+        name: String,
+        val value: String,
+        state: Int = STATE_IGNORE
+    ) : Filter.TriState(name, state)
+
     protected class GenreListFilter(genres: List<Genre>) : Filter.Group<Genre>("Genre", genres)
 
     private var genrelist: List<Genre>? = null